test(extension): move tab management tests to tab-management.spec.ts
These tests (browser_tabs new/select/close, cmd+click, window.open) are
v2-only. Grouping them in a dedicated file with a describe-level skip for
v1 avoids repeating the skip on every test.
